requirements.txt
selenium==4.23.0
beautifulsoup4==4.12.3
google-generativeai==0.4.0
pytest==8.1.1
pytest-html==4.1.1
